refactor attributes function out of world generator, try to catch invalid wait for image calls
